Update and rename WGS84 standard 2 UTM to Coords2UTM&CrossReference
In this code, the first part converts coordinates in a file from WGS84 standard to UTM. These coordinates are based in Panama City (17 in UTM). 

The second section, cross references those coordinates to polygons found in a geojson file. This is a Smithsonian Institute file found published for free in their website. If the coordinate belongs to a polygon, it will be labeled as such (Corregimiento).

Lastly, the results are exported into one excel file.